Refactor in modules
